backend/requirements.txt
torch
torchvision
redis
numpy
opencv-python-headless
wandb
Pillow